move code in main to node.rs
